% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_games.R
\name{get_top_games}
\alias{get_top_games}
\title{Get Top Games
Gets games sorted by number of current viewers on Twitch, most popular first.}
\usage{
get_top_games(first = 20, after = NULL, before = NULL)
}
\arguments{
\item{first}{Maximum number of objects to return. Maximum: 100. Default: 20.}

\item{after}{Cursor for forward pagination: tells the server where to start fetching the next set of results, in a multi-page response.}

\item{before}{Cursor for backward pagination: tells the server where to start fetching the next set of results, in a multi-page response. (Do not use, not working due to API issue last checked on 27/12/2017)}
}
\value{
A list with with "data" and a "pagination" cursor.
}
\description{
Get Top Games
Gets games sorted by number of current viewers on Twitch, most popular first.
}
\details{
If you need information about how pagination works, see the details of \link{get_streams}.
}
\references{
https://dev.twitch.tv/docs/api/reference/#get-top-games
}
