% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_users.R
\name{get_users}
\alias{get_users}
\title{Get Users
Gets information about one or more specified Twitch users. Users are identified by optional user IDs and/or login name. If neither a user ID nor a login name is specified, the user is looked up by Bearer token.}
\usage{
get_users(user_id = NULL, user_login = NULL)
}
\arguments{
\item{user_id}{User ID. Multiple user IDs can be specified. Limit: 100.}

\item{user_login}{User login name. Multiple login names can be specified. Limit: 100.}
}
\description{
Get Users
Gets information about one or more specified Twitch users. Users are identified by optional user IDs and/or login name. If neither a user ID nor a login name is specified, the user is looked up by Bearer token.
}
\references{
https://dev.twitch.tv/docs/api/reference/#get-users
}
