% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_users_follows.R
\name{get_users_follows}
\alias{get_users_follows}
\title{Get Users Follows
Gets information on follow relationships between two Twitch users. Information returned is sorted in order, most recent follow first. This can return information like "who is lirik following," "who is following lirik,” or “is user X following user Y.”}
\usage{
get_users_follows(from_id = NULL, to_id = NULL, first = 20,
  after = NULL, before = NULL)
}
\arguments{
\item{from_id}{User ID. The request returns information about users who are being followed by the from_id user.}

\item{to_id}{User ID. The request returns information about users who are following the to_id user.}

\item{first}{Maximum number of objects to return. Maximum: 100. Default: 20.}

\item{after}{Cursor for forward pagination: tells the server where to start fetching the next set of results, in a multi-page response.}

\item{before}{Cursor for backward pagination: tells the server where to start fetching the next set of results, in a multi-page response.}
}
\value{
A list with clip information.
}
\description{
Get Users Follows
Gets information on follow relationships between two Twitch users. Information returned is sorted in order, most recent follow first. This can return information like "who is lirik following," "who is following lirik,” or “is user X following user Y.”
}
