% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_videos.R
\name{get_videos}
\alias{get_videos}
\title{Get Videos
Gets video information by video ID, user ID or game ID}
\usage{
get_videos(video_id = NULL, user_id = NULL, game_id = NULL, first = 20)
}
\arguments{
\item{video_id}{Returns streams broadcast by one or more specified video IDs. You can specify up to 100 IDs.}

\item{user_id}{Returns streams broadcast by one or more specified user IDs. You can specify up to 100 IDs.}

\item{game_id}{Returns streams broadcasting a specified game ID. You can specify up to 100 IDs.}

\item{first}{Maximum number of objects to return. Maximum: 100. Default: 20.}
}
\value{
A data frame with information about the selected video ids/names.
}
\description{
Get Videos
Gets video information by video ID, user ID or game ID
}
\references{
https://dev.twitch.tv/docs/api/reference#get-videos
}
