% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetClip.R
\name{GetClip}
\alias{GetClip}
\title{Get Streams
Gets information about active streams.}
\usage{
GetClip(id)
}
\arguments{
\item{id}{ID of the clip being queried. Limit 1.}
}
\value{
A list with "data" and a "pagination" cursor.
}
\description{
Get Streams
Gets information about active streams.
}
