% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGames.R
\name{GetGames}
\alias{GetGames}
\title{Get Games
Gets game information by game ID or name.}
\usage{
GetGames(id = NULL, name = NULL)
}
\arguments{
\item{id}{Game ID. At most 100 id values can be specified.}

\item{name}{Game name. The name must be an exact match. At most 100 name values can be specified.}
}
\value{
A list with "data" and a "pagination" cursor.
}
\description{
Get Games
Gets game information by game ID or name.
}
