% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetStreams.R
\name{GetStreams}
\alias{GetStreams}
\title{Get Streams
Gets information about active streams.}
\usage{
GetStreams(first = 20, after = NULL, before = NULL, community_id = NULL,
  game_id = NULL, language = NULL, type = NULL, user_id = NULL,
  user_login = NULL)
}
\arguments{
\item{first}{Maximum number of objects to return. Maximum: 100. Default: 20.}

\item{after}{Cursor for forward pagination: tells the server where to start fetching the next set of results, in a multi-page response.}

\item{before}{Cursor for backward pagination: tells the server where to start fetching the next set of results, in a multi-page response. (Do not use, not working due to API issue last checked on 27/12/2017}

\item{community_id}{Returns streams in a specified community ID. You can specify up to 100 IDs.}

\item{game_id}{Returns streams broadcasting a specified game ID. You can specify up to 100 IDs.}

\item{language}{Stream language. You can specify up to 100 languages.}

\item{type}{Stream type: "all", "live", "vodcast". Default: "all".}

\item{user_id}{Returns streams broadcast by one or more specified user IDs. You can specify up to 100 IDs.}

\item{user_login}{Returns streams broadcast by one or more specified user login names. You can specify up to 100 names.}

\item{dataOnly}{If TRUE, removes pagination element.}
}
\value{
A list with "data" and a "pagination" cursor.
}
\description{
Get Streams
Gets information about active streams.
}
\details{
For multiple query entries, use one string with comma-separated entries. Example: language="pt-br,en-us". Streams are returned sorted by number of current viewers, in descending order. Across multiple pages of results, there may be duplicate or missing streams, as viewers join and leave streams.
}
