% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTopGames.R
\name{GetTopGames}
\alias{GetTopGames}
\title{Get Top Games
Gets games sorted by number of current viewers on Twitch, most popular first.}
\usage{
GetTopGames(clientID, first = 20, after = NULL, before = NULL)
}
\arguments{
\item{first}{Maximum number of objects to return. Maximum: 100. Default: 20.}

\item{after}{Cursor for forward pagination: tells the server where to start fetching the next set of results, in a multi-page response.}

\item{before}{Cursor for backward pagination: tells the server where to start fetching the next set of results, in a multi-page response. (Do not use, not working due to API issue last checked on 27/12/2017}

\item{ClientID}{your clientID}
}
\description{
Get Top Games
Gets games sorted by number of current viewers on Twitch, most popular first.
}
