% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{twitch_auth}
\alias{twitch_auth}
\title{Authenctication
Sets up your Client-ID.}
\usage{
twitch_auth(client_id)
}
\arguments{
\item{clientID}{your client-ID}
}
\description{
Authenctication
Sets up your Client-ID.
}
