% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{twitch_auth_forget}
\alias{twitch_auth_forget}
\title{Reset Authenctication}
\usage{
twitch_auth_forget()
}
\description{
Forgets Client-ID.
}
\details{
Use this to forget the client-ID previously set, then set your new client-ID with \code{\link{twitch_auth}}
}
