% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitchAuth.R
\name{twitchAuth}
\alias{twitchAuth}
\title{Authenctication
Sets up your Client-ID.}
\usage{
twitchAuth(clientID)
}
\arguments{
\item{clientID}{your client-ID}
}
\description{
Authenctication
Sets up your Client-ID.
}
