% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clip.R
\name{get_clip}
\alias{get_clip}
\title{Get Clip
Gets information about clip.}
\usage{
get_clip(id)
}
\arguments{
\item{id}{ID of the clip being queried. Limit 1.}
}
\value{
A list with clip information.
}
\description{
Get Clip
Gets information about clip.
}
